Refactor request paths
