+++ src/main/kotlin/ru/neexol/routes/api/notes/authorsRoute.kt
@@ -1,4 +1,4 @@
-package ru.neexol.routes.api
+package ru.neexol.routes.api.notes
 
 import io.ktor.server.application.*
 import io.ktor.server.response.*
