+++ src/main/kotlin/ru/neexol/routes/api/schedule/updateRoute.kt
@@ -1,4 +1,4 @@
-package ru.neexol.routes.api
+package ru.neexol.routes.api.schedule
 
 import io.ktor.server.application.*
 import io.ktor.server.response.*
