+++ src/main/kotlin/ru/neexol/routes/apiRoute.kt
@@ -1,13 +1,10 @@
 package ru.neexol.routes
 
 import io.ktor.server.routing.*
-import ru.neexol.routes.api.*
+import ru.neexol.routes.api.notes.notesRoute
+import ru.neexol.routes.api.schedule.scheduleRoute
 
 fun Route.apiRoute() {
-    timesRoute()
-    updateRoute()
-    checksumRoute()
     scheduleRoute()
     notesRoute()
-    authorsRoute()
 }
